+++ src/ClassicUO.Client/Game/Scenes/GameSceneInputHandler.cs
@@ -685,7 +685,6 @@ namespace ClassicUO.Game.Scenes
                     case CursorTarget.Object:
                     case CursorTarget.MultiPlacement when _world.CustomHouseManager == null:
                     case CursorTarget.CallbackTarget:
-
                         {
                             BaseGameObject obj = lastObj;